Condense all to client.go
