--- exporters/otlp/otlpmetric/internal/otest/collector.go
@@ -1,58 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//go:build go1.18
-// +build go1.18
-
-package otest // import "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal/otest"
-
-import (
-	"sync"
-
-	cpb "go.opentelemetry.io/proto/otlp/collector/metrics/v1"
-	mpb "go.opentelemetry.io/proto/otlp/metrics/v1"
-)
-
-// Collector is the collection target a Client sends metric uploads to.
-type Collector interface {
-	Collect() *Storage
-}
-
-// Storage stores uploaded OTLP metric data in their proto form.
-type Storage struct {
-	dataMu sync.Mutex
-	data   []*mpb.ResourceMetrics
-}
-
-// NewStorage returns a configure storage ready to store recieved requests.
-func NewStorage() *Storage {
-	return &Storage{}
-}
-
-// Add adds the request to the Storage.
-func (s *Storage) Add(request *cpb.ExportMetricsServiceRequest) {
-	s.dataMu.Lock()
-	defer s.dataMu.Unlock()
-	s.data = append(s.data, request.ResourceMetrics...)
-}
-
-// dump returns all added ResourceMetrics and clears the storage.
-func (s *Storage) dump() []*mpb.ResourceMetrics {
-	s.dataMu.Lock()
-	defer s.dataMu.Unlock()
-
-	var data []*mpb.ResourceMetrics
-	data, s.data = s.data, []*mpb.ResourceMetrics{}
-	return data
-}
